Rewrite AIPs

+++ src/main/java/swst/application/repositories/OrdersRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import swst.application.entities.Orders;
 
-public interface OrdersRepository extends JpaRepository<Orders, Integer> {
+public interface OrdersRepository extends JpaRepository<Orders, Long> {
 
 }
